[refactor] Create taichi_isolated_core that is free from pybind11
