+++ cmake/TaichiTests.cmake
@@ -16,7 +16,7 @@ include_directories(
 )
 
 add_executable(${TESTS_NAME} ${TAICHI_TESTS_SOURCE})
-target_link_libraries(${TESTS_NAME} taichi_testable_lib)
+target_link_libraries(${TESTS_NAME} taichi_isolated_core)
 target_link_libraries(${TESTS_NAME} gtest_main)
 
 add_test(NAME ${TESTS_NAME} COMMAND ${TESTS_NAME})
