+++ taichi/backends/cuda/detect_cuda.h
@@ -1,0 +1,5 @@
+#pragma once
+
+namespace taichi {
+bool is_cuda_api_available();
+}  // namespace taichi
