+++ taichi/program/py_print_buffer.cpp
@@ -1,0 +1,7 @@
+#include "taichi/program/py_print_buffer.h"
+
+namespace taichi {
+
+PythonPrintBuffer py_cout;
+
+}  // namespace taichi
